Separate request/response schemas
